PSculpt: Adjust Brush - Hacky WIP "Rotate" mode
This commit introduces some WIP tweaks to the behaviour of the adjust brush
when "shift" is applied. Instead of switching to translation mode, it will
now switch to an alternative rotation behaviour (based on the way that
"normal" rotate works).

Currently this is a bit messed up, and only works if you go "up" the chain
(i.e. start from the tip of a bone and draw in to its head to get it to
rotate). This is actually the opposite of what should be happening, so it
feels quite weird/nasty to use. I'm not yet sure where it's coming from though.