PSculpt: Clamp distances used in the influence calculations
It's possible that a bone fits inside the brush while its "distance" is judged
to be outside the brush (e.g. a single bone tip on a long bone). To try and
prevent "runaways" (i.e. compounding influence due to the distance function
going below zero, and rebounding), we now clamp the distance so that that
cannot happen.