PSculpt: "Rotate" brush mode (Adjust + Shift) works now
This commit includes various changes made in a bid to get this working.
Some of these may not be necessary though.

The key measure here seems to be the step where the rotation angle gets
inverted. This is because bones were always getting rotated in the opposite
direction to what we were expecting.