PSketch: Location offsets are now applied correctly
Instead of trying to apply offsets, we now just snap the bone location to
the desired stroke coordinates directly. These coordinates are already in
pose space, so we just need to apply a pose space -> bone space conversion
(using the method used for the posebone snapping).

Also, when applying locations, we temporarily lock all locations so that
it doesn't try to rotate around anything it shouldn't.