PSketch: First attempt to make it not rotate set the rotation going the wrong way
Instead of using the "compatible rotation" method for using the smallest difference
with the previous rotation state of the bone, let's just do nothing. That's because,
sometimes the smallest rotational difference ends up being to go straight through
the chain (which is wrong).

Maybe better is to check that it doesn't go past +/- 180, or to make it use the
previous frame's pose as reference instead?

More testing needed!