PSketch: Hacky first attempt at limiting deforms to x or z axes
This is a hacky feature to make it so that you don't have to align your view
precisely with the bone chain in a particular way in order to sketch the desired
pose. Instead, you can make it only apply the x or z components of the implied
transforms.

The current implementation does it by temporarily applying the rotation lock
flags to the appropriate channels when performing the rotations. It works ok
in many situations, but when extreme curves are used from an off angle,
some of the "power" of the stroke does get lost, meaning that it doesn't
curve quite as much as would be expected (BAD!)  Maybe we'll need to implement
some fancier projection stuff at a more fundamental level instead...